+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   [#555](https://github.com/hyperledger-labs/fablo/pull/555)
 * Include and test Sample Go chaincode in samples/chaincodes/chaincode-kv-go
   [#569](https://github.com/hyperledger-labs/fablo/pull/569)
+* Export network topology with Mermaid 
+  [#565](https://github.com/hyperledger-labs/fablo/pull/565)
 
 
 ## 2.2.0